# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  github_com_Ulpio_vergo_internal_domain_org.Organization:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owner_user_id:
        type: string
    type: object
...
  internal_http_handlers.ErrorDetailResponse:
    description: Error response with detail
    properties:
      detail:
        example: duplicate key
        type: string
      error:
        example: create_fail
        type: string
    type: object
  internal_http_handlers.ErrorResponse:
    description: Standard error response
    properties:
      error:
        example: not_found
        type: string
    type: object
...
  internal_http_handlers.createOrgIn:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  internal_http_handlers.memberIn:
    properties:
      role:
        description: owner|admin|member
        type: string
      user_id:
        type: string
    required:
    - role
    - user_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Multi-tenant SaaS boilerplate API with JWT auth, RBAC, and S3 storage.
  title: Vergo API
  version: 0.1.0
paths:
...
  /orgs:
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.createOrgIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Ulpio_vergo_internal_domain_org.Organization'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create organization
      tags:
      - Organizations
  /orgs/{id}:
    delete:
      parameters:
      - description: Organization ID
        in: header
        name: X-Org-ID
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorDetailResponse'
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - Organizations
    get:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Ulpio_vergo_internal_domain_org.Organization'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get organization
      tags:
      - Organizations
  /orgs/{id}/members:
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: header
        name: X-Org-ID
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Member details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.memberIn'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add member
      tags:
      - Organizations
  /orgs/{id}/members/{userId}:
    delete:
      parameters:
      - description: Organization ID
        in: header
        name: X-Org-ID
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorDetailResponse'
      security:
      - BearerAuth: []
      summary: Remove member
      tags:
      - Organizations
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: header
        name: X-Org-ID
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.memberIn'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorDetailResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update member role
      tags:
      - Organizations
...
securityDefinitions:
  BearerAuth:
    description: Bearer token (e.g. "Bearer eyJhbG...")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"